edit code location and design according to review
