+++ packages/salesforcedx-vscode-lwc/src/index.ts
@@ -27,8 +27,8 @@ import {
   forceLightningLwcStart,
   forceLightningLwcStop
 } from './commands';
-import { addJsMetaSetting, createLWCResource } from './commands/createResources';
 import { ESLINT_NODEPATH_CONFIG, LWC_EXTENSION_NAME } from './constants';
+import { metaSupport } from './metasupport';
 import { DevServerService } from './service/devServerService';
 import { telemetryService } from './telemetry';
 import {
@@ -100,8 +100,7 @@ export async function activate(context: ExtensionContext) {
   startLWCLanguageServer(context);
 
   // Creates resources for js-meta.xml to work
-  createLWCResource();
-  addJsMetaSetting();
+  metaSupport.getMetaSupport();
 
   if (workspaceType === lspCommon.WorkspaceType.SFDX) {
     // We no longer want to manage the eslint.nodePath. Remove any previous configuration of the nodepath
