+++ packages/salesforcedx-vscode-lwc/src/metasupport/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2018, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import { MetaSupport } from './metaSupport';
+
+export const metaSupport = MetaSupport.initializeSupport();
